Move shared logic out of route modules
