+++ apps/web/utils/ai/assistant/chat-context-validation.ts
