+++ apps/web/utils/ai/assistant/chat.ts
@@ -1,6 +1,6 @@
 import type { JSONValue, ModelMessage } from "ai";
 import type { Logger } from "@/utils/logger";
-import type { MessageContext } from "@/app/api/chat/validation";
+import type { MessageContext } from "@/utils/ai/assistant/chat-context-validation";
 import { stringifyEmail } from "@/utils/stringify-email";
 import { getEmailForLLM } from "@/utils/get-email-from-message";
 import type { ParsedMessage } from "@/utils/types";
